refactor: move `handle` functions out
