+++ src/components/PartCard/SearchInfo/SearchInfo.tsx
@@ -1,14 +1,7 @@
 import { format } from 'date-fns';
-import { useContext, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 
-import { MutateItemList } from '../../../services/apiTypes';
-
-import { useAddItemsToList } from '../../../services/hooks/items/useAddItemsToList';
-import { useRemoveItemsFromList } from '../../../services/hooks/items/useRemoveItemsFromList';
-
-import UmAppContext from '../../../contexts/UmAppContext';
-import { useGetListById } from '../../../services/hooks/list/useGetListById';
+import { usePartActions } from '../../../services/hooks/list/usePartCardActions';
 import CustomDialog from '../../CustomDialog/CustomDialog';
 import { StyledAddIcon, StyledInfoIcon, StyledRemoveIcon } from '../../ListCard/styles';
 import { PartCardProps } from '../PartCard';
@@ -16,61 +9,19 @@ import { KeyWords } from '../styles';
 import { DescriptionParagraph, FirstInfoBox, InfoP, SecondInfoBox, ThirdInfoBox } from './styles';
 
 export const SearchInfo = ({ part, icon }: PartCardProps) => {
-    const { setSnackbarText, setSnackbarSeverity } = useContext(UmAppContext);
-
-    const [open, setOpen] = useState(false);
-    const [alreadyAdded, setAlreadyAdded] = useState(false);
     const { listId } = useParams();
     const navigate = useNavigate();
-    const { data: list } = useGetListById(listId!);
-    const { mutate: mutateAddItemToList, isSuccess: addItemSuccess } = useAddItemsToList();
-    const { mutate: mutateRemoveItemFromList } = useRemoveItemsFromList();
 
-    const handleAdd = (e: React.MouseEvent, ids: MutateItemList) => {
-        e.stopPropagation();
-        const alreadyAdded = list?.items.some((item) => item.id === part.id);
-        if (alreadyAdded) {
-            {
-                setAlreadyAdded(true);
-            }
-            setSnackbarSeverity('error');
-            setSnackbarText('already in list');
-        }
-        mutateAddItemToList(ids, {
-            onSuccess: (data) => {
-                if (alreadyAdded) return;
-                setSnackbarText(`${part.wpId} was added`);
+    const {
+        handleAdd,
+        handleDelete,
+        handleClickOpen,
+        handleClose,
+        open,
+        alreadyAdded,
+        addItemSuccess,
+    } = usePartActions({ part, icon }, listId);
 
-                if (data.status >= 400) {
-                    setSnackbarSeverity('error');
-                    setSnackbarText(`${data.statusText}, please try again.`);
-                }
-            },
-        });
-        handleClose(e);
-    };
-    const handleDelete = (e: React.MouseEvent, ids: MutateItemList) => {
-        e.stopPropagation();
-        mutateRemoveItemFromList(ids, {
-            onSuccess: (removeData) => {
-                setSnackbarText(`${part.wpId} was removed`);
-
-                if (removeData.status >= 400) {
-                    setSnackbarSeverity('error');
-                    setSnackbarText(`${removeData.statusText}, please try again.`);
-                }
-            },
-        });
-        handleClose(e);
-    };
-    const handleClickOpen = (e: React.MouseEvent) => {
-        e.stopPropagation();
-        setOpen(true);
-    };
-    const handleClose = (e: React.MouseEvent) => {
-        e.stopPropagation();
-        setOpen(false);
-    };
     return (
         <>
             <FirstInfoBox>
